+++ include/nsmb.h
@@ -1,0 +1,21 @@
+#ifndef NSMB_H_
+#define NSMB_H_
+
+#include "nsmb/actor.h"
+#include "nsmb/collider.h"
+#include "nsmb/file.h"
+#include "nsmb/font.h"
+#include "nsmb/global.h"
+#include "nsmb/math.h"
+#include "nsmb/memory.h"
+#include "nsmb/model.h"
+#include "nsmb/oam.h"
+#include "nsmb/particle.h"
+#include "nsmb/player.h"
+#include "nsmb/save.h"
+#include "nsmb/sound.h"
+#include "nsmb/util.h"
+#include "nsmb/vector.h"
+#include "nsmb/worldmap.h"
+
+#endif // !NSMB_H_
